% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.svd.data}
\alias{as.svd.data}
\title{Converts an object to an svd.data object}
\usage{
as.svd.data(obj, ...)
}
\arguments{
\item{obj}{the object to be converted}

\item{...}{additional params}
}
\description{
Converts an object to an svd.data object
}

