% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.svd.plate}
\alias{as.svd.plate}
\title{Convert to an plate object}
\usage{
as.svd.plate(obj, ...)
}
\arguments{
\item{obj}{objject to be converted}

\item{...}{additional params}
}
\description{
Convert to an plate object
}

