% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_statistic.R
\name{chisq.statistic}
\alias{chisq.statistic}
\title{Calculate statistics based on the chi-square-distribution to analyse the data.}
\usage{
chisq.statistic(obj, padjust = c("BH", "bonferroni"), ...)
}
\arguments{
\item{obj}{the data to be analysed}

\item{padjust}{multiple testing correction method}

\item{...}{additional params}
}
\description{
For this you should use a standardization before.
}

