% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_knockout_analysed.R
\docType{class}
\name{ChisqStatisticAnalysis-class}
\alias{ChisqStatisticAnalysis-class}
\title{Data wrapper for analysed knockout data using a standard hypothesis test}
\description{
Class \code{knockout.chisqstatistic.analysed} is a wrapper for a
  \code{data.table} object containing the knockout data
}
\section{Slots}{

\describe{
\item{\code{.gene.hits}}{prioritized genes}
}}

