% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correct}
\alias{correct}
\title{Off-target correct and rank by hit magnitude}
\usage{
correct(obj, do.pooled = F, ...)
}
\arguments{
\item{obj}{a summarized data.table}

\item{do.pooled}{boolean flag whether pooled libraries should also be corrected}

\item{...}{additional arguments}
}
\description{
Off-target correct and rank by hit magnitude
}

