% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_correlation.R
\name{correlation}
\alias{correlation}
\title{Calculate the correlation between two data-sets}
\usage{
correlation(x, y, method = c("pearson", "kendall", "spearman"), ...)
}
\arguments{
\item{x}{a random object}

\item{y}{a random object of the same type (optional)}

\item{method}{a character string indicating which correlation coefficient
(\emph{pearson}, \emph{kendall} or \emph{spearman})}

\item{...}{additional parameters}
}
\description{
Calculate the correlation between two data-sets
}
