% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_getters.R, R/class_analysed.R,
%   R/class_data.R
\docType{methods}
\name{dataSet}
\alias{dataSet}
\alias{dataSet,AbstractAnalysedPerturbationData-method}
\alias{dataSet,PerturbationData-method}
\title{Getter for a data set}
\usage{
dataSet(obj)

\S4method{dataSet}{AbstractAnalysedPerturbationData}(obj)

\S4method{dataSet}{PerturbationData}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the underlying data}
}
\value{
returns a \code{data.table}.
}
\description{
Returns the data set that underlies an S4 wrapper class
 as \code{data.table}.
}
\examples{
 data(rnaiscreen)
 dataSet(rnaiscreen)
}
