% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_getters.R, R/class_data.R
\docType{methods}
\name{dataType}
\alias{dataType}
\alias{dataType,PerturbationData-method}
\title{Getter for a data set type}
\usage{
dataType(obj)

\S4method{dataType}{PerturbationData}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the underlying data type}
}
\value{
returns a \code{character}.
}
\description{
Returns the type of data set that underlies an S4 wrapper class
 as \code{character}. We distinguish between different data types, like raw
 or normalized, in order to keep track of which methods can be applied to
 the S4 wrapper.
}
