% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_knockout_analysed.R
\docType{class}
\name{DiffusionAnalysis-class}
\alias{DiffusionAnalysis-class}
\title{Data wrapper for analysed knockout data using network diffusion}
\description{
Class \code{knockout.analysed.diffusion} is a wrapper for a
  \code{data.table} object containing the knockout data
}
\section{Slots}{

\describe{
\item{\code{.is.bootstrapped}}{boolean whether bootstrapping has been done or not}
}}

