% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysed.R
\docType{class}
\name{DiffusionAnalysis-class}
\alias{DiffusionAnalysis-class}
\title{Data wrapper for analysed perturbation data using network diffusion}
\description{
Class \code{perturbation.diffusion.analysed} is a wrapper for a
  \code{data.table} object containing the perturbation data
}
\section{Slots}{

\describe{
\item{\code{.intitial.model}}{the model that was provided for analysis}

\item{\code{.graph}}{an igraph object that served for the diffusion process}
}}

