% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{effect.matrices}
\alias{effect.matrices}
\title{Calculcate the I-dont-know-how-to-call-them}
\usage{
effect.matrices(obj, ...)
}
\arguments{
\item{obj}{the object to calculate the ffect matrices for}

\item{...}{additional parameters}
}
\description{
Calculcate the I-dont-know-how-to-call-them
}

