% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_filter.R
\name{filter}
\alias{filter}
\title{Filter the rows of a knockdown data set}
\usage{
filter(obj, ...)
}
\arguments{
\item{obj}{the object to be filtered}

\item{...}{additional parameters}
}
\value{
returns an object of the same type filtered by some criterion
}
\description{
Takes a knockdown data set and filters the rows by some
 criterion. The filtered object will have the same type as the previous
 object.
}
\examples{
 data(rnaiscreen)
 flt.dat <- filter(rnaiscreen, Virus=="V1")
}
