% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_getters.R, R/class_analysed.R
\docType{methods}
\name{geneEffects}
\alias{geneEffects}
\alias{geneEffects,HMAnalysedPerturbationData-method}
\title{Getter for the complete list of genes and their effect sizes}
\usage{
geneEffects(obj)

\S4method{geneEffects}{HMAnalysedPerturbationData}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the underlying gene}
}
\value{
returns a \code{list}.
}
\description{
Returns a \code{data.table} containing the list of genes that
have been used in the study along with their estimated effect sizes.
}
