% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_getters.R, R/class_analysed.R
\docType{methods}
\name{geneHits}
\alias{geneHits}
\alias{geneHits,HMAnalysedPerturbationData-method}
\title{Getter for identified essential genes}
\usage{
geneHits(obj)

\S4method{geneHits}{HMAnalysedPerturbationData}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the underlying gene hits}
}
\value{
returns a \code{data.table}.
}
\description{
Returns a \code{data.table} containing the genes that have been
 identified as essential genes in a perturbation screen.
}
\examples{
 data(rnaiscreen)
 ft <- hm(rnaiscreen)
 geneHits(ft)
}
