% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_getters.R, R/class_analysed.R
\docType{methods}
\name{graph}
\alias{graph}
\alias{graph,NetworkAnalysedPerturbationData-method}
\title{Getter for graph used for network diffusion}
\usage{
graph(obj)

\S4method{graph}{NetworkAnalysedPerturbationData}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the underlying graph}
}
\value{
returns a \code{igraph} object
}
\description{
Returns the graph that has been used for analysis using network
 diffusion.
}
\examples{
 data(rnaiscreen)
 graph.file <- system.file("extdata", "graph_file.tsv",
                            package = "perturbatr")
 ft <- hm(rnaiscreen, effect.size=0.01)
 diffu <- diffuse(ft, path=graph.file, r=1)
 graph(diffu)
}
