% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysed.R
\docType{class}
\name{HMAnalysis-class}
\alias{HMAnalysis-class}
\title{Data wrapper for analysed perturbation data using a hierarchical model}
\description{
Class \code{perturbation.hm.analysed} is a wrapper for a
  \code{data.table} object containing the perturbation data
}
\section{Slots}{

\describe{
\item{\code{.gene.effects}}{the estimated effect sizes for genes}

\item{\code{.nested.gene.effects}}{the estimated effect sizes for genes on a
viral level}

\item{\code{.gene.hits}}{prioritized genes}

\item{\code{.neted.gene.hits}}{nested prioritized genes}

\item{\code{.model.fit}}{the fitted model}
}}

