% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysed.R
\docType{class}
\name{HMAnalysedPerturbationData-class}
\alias{HMAnalysedPerturbationData-class}
\title{Data wrapper for analysed perturbation data using a hierarchical model}
\description{
Class \code{HMAnalysedPerturbationData} is a wrapper for
 various objects of an analysis of a perturbation experiment done
 using a hierarchical model. \code{HMAnalysedPerturbationData} exposes
 getters for its members of the same name.
 However, we do not provide setters, because the data should be treated as
 constant once set.
}
\section{Slots}{

\describe{
\item{\code{geneHits}}{prioritized genes}

\item{\code{nestedGeneEffects}}{the estimated effect sizes for genes on a
viral level}

\item{\code{nestedGeneHits}}{nested prioritized genes}

\item{\code{modelFit}}{the fitted model}
}}

