% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysed.R
\docType{class}
\name{HyperAnalysis-class}
\alias{HyperAnalysis-class}
\title{Data wrapper for analysed perturbation data using a standard hypothesis test}
\description{
Class \code{perturbation.hyper.analysed} is a wrapper for a
  \code{data.table} object containing the perturbation data
}
\section{Slots}{

\describe{
\item{\code{.gene.hits}}{prioritized genes}
}}

