% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_hyperstatistic.R
\docType{methods}
\name{hyper.statistic}
\alias{hyper.statistic}
\alias{hyper.statistic,perturbation.normalized.data-method}
\alias{hyper.statistic,perturbation.data-method}
\title{Calculate statistics based on the hypergeometric-distribution to
 analyse the data.}
\usage{
hyper.statistic(obj, padjust = c("BH", "bonferroni"),
  summ.method = c("mean", "median"), level = c("gene", "sirna"),
  do.summarization = FALSE, hit.ratio = 0.5, effect.size = 0,
  pval.threshold = 0.05, qval.threshold = 1)

\S4method{hyper.statistic}{perturbation.normalized.data}(obj,
  padjust = c("BH", "bonferroni"), summ.method = c("mean", "median"),
  level = c("gene", "sirna"), do.summarization = FALSE, hit.ratio = 0.5,
  effect.size = 0, pval.threshold = 0.05, qval.threshold = 1)
}
\arguments{
\item{obj}{the data to be analysed}

\item{padjust}{multiple testing correction method}

\item{summ.method}{summarize single siRNAs using mean or median}

\item{level}{do hypergeometric test on gene level or siRNA level.
If level=sirna is chosen, multiple replicates have to be given.
If level=gene we use all the sirnas for a gene and optionally summarize
siRNAs over replicate level.}

\item{do.summarization}{boolean flag whether sirnas should be summarized
if level=gene is chosen}

\item{hit.ratio}{ratio of succesful hits such that a gene is considered a
'hit'}

\item{effect.size}{the mean effect size used for hit prioritization}

\item{pval.threshold}{p-value threshold when a siRNA should be considered a
'hit'}

\item{qval.threshold}{q-value threshold when a siRNA should be considered a
'hit'}
}
\value{
returns a \code{perturbation.hyper.analysed} object
}
\description{
TODO
}
\examples{
 data(rnaiscreen)
 v1.dat <- perturbatr::filter(rnaiscreen, Condition=="V1")
 v1.data.norm <- perturbatr::preprocess(v1.dat)
 v1.res <- hyper.statistic(v1.data.norm)
}
