% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_graph.R
\name{induced.subgraph}
\alias{induced.subgraph}
\title{Get the neighbors of a node including all edges between them}
\usage{
induced.subgraph(gene, graph)
}
\arguments{
\item{gene}{the gene for which the neighborhood is searched}

\item{graph}{the graph}
}
\description{
Get the neighbors of a node including all edges between them
}
