% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_getters.R, R/class_analysed.R
\docType{methods}
\name{inference}
\alias{inference}
\alias{inference,AbstractAnalysedPerturbationData-method}
\title{Getter for inference used for analysis of perturbation data}
\usage{
inference(obj)

\S4method{inference}{AbstractAnalysedPerturbationData}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the underlying inference}
}
\value{
returns a \code{character}.
}
\description{
Returns the inference used in the analysis of a perturbation
 analysis. This can for instance be a standard t-test or a hierarchical
 model.
}
\examples{
 data(rnaiscreen)
 ft <- hm(rnaiscreen)
 inference(ft)
}
