% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_getters.R, R/class_analysed.R
\docType{methods}
\name{isBootstrapped}
\alias{isBootstrapped}
\alias{isBootstrapped,AbstractAnalysedPerturbationData-method}
\title{Getter for boolean if boostrapping was used}
\usage{
isBootstrapped(obj)

\S4method{isBootstrapped}{AbstractAnalysedPerturbationData}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the boolean}
}
\value{
returns a \code{boolean}.
}
\description{
Returns a boolean if for the analysed object bootstrapping was
 used to create confidence intervals.
}
