% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_knockout_analysed.R
\docType{class}
\name{kNN-DiffusionAnalysis-class}
\alias{kNN-DiffusionAnalysis-class}
\title{Data wrapper for analysed knockout data using nearest neighbors}
\description{
Class \code{knockout.mrw.diffusion.analysed} is a wrapper for a
  \code{data.table} object containing the knockout data
}
\section{Slots}{

\describe{
\item{\code{.neighbors}}{a \code{data.table} with found neighbors for every start
gene}
}}

