% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockdown_package.R
\docType{package}
\name{knockdown-package}
\alias{knockdown-package}
\title{knockdown}
\description{
Analysis of high-throughput gene perturbation screens.
}
\author{
Simon Dirmeier | Computational Biology Group,
 ETH ZURICH | \email{simon.dirmeier@bsse.ethz.ch}
}
\keyword{package}
