% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_knockdown_analysed.R
\docType{class}
\name{KnockdownAnalysis-class}
\alias{KnockdownAnalysis-class}
\title{Data wrapper for analysed knockdown data}
\description{
Abstract class \code{knockdown.analysed} is a wrapper for a
  \code{data.table} object
containing the knockdown data
}
\section{Slots}{

\describe{
\item{\code{.data}}{the knockdown data-set}

\item{\code{.inference}}{the method for inferenced that has been used}

\item{\code{.is.bootstrapped}}{boolean whether bootstrap intervals have been}

\item{\code{.params}}{list of some used parameters
created or not}
}}

