% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_knockdown_data.R
\docType{class}
\name{Normalized-data}
\alias{Normalized-data}
\title{Data wrapper for normalized knockdown data}
\description{
Class \code{knockdown.normalized.data} is a wrapper for a
\code{data.table} object containing the normalized data-set
}
