% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockout-package.R
\docType{package}
\name{knockout-package}
\alias{knockout-package}
\title{knockout}
\description{
Analysis of high-throughput gene perturbation screens.
}
\author{
Simon Dirmeier | Computational Biology Group, ETH ZURICH | \email{simon.dirmeier@bsse.ethz.ch}
}
\keyword{package}

