% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{knockout.data-class}
\alias{knockout.data-class}
\alias{knockout.data}
\title{Data wrapper for knockout data}
\description{
Class \code{knockout.data} is a wrapper for a \code{data.table} object
containing the knockout data
}
\section{Slots}{

\describe{
\item{\code{.data}}{the knockout data-set}
}}

