% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-knockout_analysed.R
\docType{class}
\name{Analysed knockout data}
\alias{Analysed knockout data}
\alias{Analysed knockout data}
\title{Data wrapper for analysed knockout data using an LMM}
\description{
Abstract class \code{knockout.data} is a wrapper for a
  \code{data.table} object
containing the knockout data

Class \code{knockout.data} is a wrapper for a
  \code{data.table} object
containing the knockout data
}
\section{Slots}{

\describe{
\item{\code{.data}}{the knockout data-set}

\item{\code{.inference}}{the method for inferenced that has been used}

\item{\code{.is.bootstrapped}}{boolean whether bootstrapping has been done or not}
}}

