% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_knockout_analysed.R
\docType{class}
\name{KnockoutAnalysis-class}
\alias{KnockoutAnalysis-class}
\title{Data wrapper for analysed knockout data}
\description{
Abstract class \code{knockout.analysed} is a wrapper for a
  \code{data.table} object
containing the knockout data
}
\section{Slots}{

\describe{
\item{\code{.data}}{the knockout data-set}

\item{\code{.inference}}{the method for inferenced that has been used}
}}

