% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_knockout_data.R
\docType{class}
\name{LMM-data}
\alias{LMM-data}
\title{Data wrapper for linear-mixed-model knockout data}
\description{
Class \code{knockout.lmm.data} is a wrapper for a
\code{data.table} object containing the normalized data-set
}
