% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_knockout_data.R
\docType{class}
\name{KnockoutData-class}
\alias{KnockoutData-class}
\title{Data wrapper for knockout data}
\description{
Class \code{knockout.data} is a wrapper for a
\code{data.table} object containing the knockout data
}
\section{Slots}{

\describe{
\item{\code{.data}}{the knockout data-set}
}}

