% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_lmm.R
\docType{methods}
\name{lmm}
\alias{lmm}
\alias{lmm,knockout.normalized.data-method}
\alias{lmm,knockout.lmm.data-method}
\title{Fit an LMM to the data and calculate local false discovery rates.}
\usage{
lmm(obj, drop = T, weights = NULL, rel.mat.path = NULL,
  bootstrap.cnt = 0, ignore = 1, effect.size = 0.05,
  qval.threshold = 0.2)

\S4method{lmm}{knockout.normalized.data}(obj, drop = T, weights = NULL,
  rel.mat.path = NULL, bootstrap.cnt = 0, ignore = 1,
  effect.size = 0.05, qval.threshold = 0.2)

\S4method{lmm}{knockout.lmm.data}(obj, drop = T, weights = NULL,
  rel.mat.path = NULL, bootstrap.cnt = 0, ignore = 1,
  effect.size = 0.01, qval.threshold = 0.2)
}
\arguments{
\item{obj}{an svd.data object}

\item{drop}{boolean flag if all entries should be dropped that are not
found in every virus}

\item{weights}{a list of weights}

\item{rel.mat.path}{the (optional) path to a target relation matrix that is
going to be used for}

\item{bootstrap.cnt}{the number of bootstrap runs you want to do in order
to estimate a significance level for the gene effects}

\item{ignore}{ignore siRNAs that have been seen only once per group}

\item{effect.size}{the relative effect size used for hit prioritization}

\item{qval.threshold}{the q-value threshold used for hit prioritization
if bootstrap.cnt is set}
}
\description{
\code{lmm} TODO
}
