% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_knockout_analysed.R
\docType{class}
\name{LMMAnalysis-class}
\alias{LMMAnalysis-class}
\title{Data wrapper for analysed knockout data using an LMM}
\description{
Class \code{knockout.analysed.lmm} is a wrapper for a
  \code{data.table} object
containing the knockout data
}
\section{Slots}{

\describe{
\item{\code{.is.bootstrapped}}{boolean whether bootstrapping has been done or not}
}}

