% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model.data.lmm}
\alias{model.data.lmm}
\title{Create model data for an LMM}
\usage{
model.data.lmm(obj, drop = T, ignore = 1, weights = NULL,
  rel.mat.path = NULL)
}
\arguments{
\item{obj}{an object for which LMM model.data is created}

\item{drop}{decide if genes that are not found in every virus should be dropped}

\item{ignore}{ignore siRNAS that are only found \code{ignore} many times}

\item{weights}{weights to set for the siRNAs}

\item{rel.mat.path}{target-relation matrix (TODO)}
}
\description{
Create model data for an LMM
}
