% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_getters.R, R/class_analysed.R
\docType{methods}
\name{modelFit}
\alias{modelFit}
\alias{modelFit,HMAnalysedPerturbationData-method}
\title{Getter for model fit}
\usage{
modelFit(obj)

\S4method{modelFit}{HMAnalysedPerturbationData}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the underlying fit}
}
\value{
returns a \code{list}
}
\description{
Returns the mixed effects model fit from the analysis using a
 hierarchical model.
}
\examples{
 data(rnaiscreen)
 hm.fit <- hm(rnaiscreen)
 ft <- modelFit(hm.fit)
}
