% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_knockout_analysed.R
\docType{class}
\name{MRW-DiffusionAnalysis-class}
\alias{MRW-DiffusionAnalysis-class}
\title{Data wrapper for analysed knockout data using Markov random walks
 with restarts}
\description{
Class \code{knockout.mrw.diffusion.analysed} is a wrapper for a
  \code{data.table} object containing the knockout data
}
