% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_getters.R, R/class_analysed.R
\docType{methods}
\name{nestedGeneEffects}
\alias{nestedGeneEffects}
\alias{nestedGeneEffects,HMAnalysedPerturbationData-method}
\title{Getter for the completed list of nested gene effects}
\usage{
nestedGeneEffects(obj)

\S4method{nestedGeneEffects}{HMAnalysedPerturbationData}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the underlying effects}
}
\value{
returns a \code{tibble}.
}
\description{
Returns a \code{tibble} containing the list of nested genes
 effects, i.e. a table of nested cluster effects.
}
\examples{
 data(rnaiscreen)
 ft <- hm(rnaiscreen)
 nestedGeneEffects(ft)
}
