% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysed.R
\docType{class}
\name{NetworkAnalysedPerturbationData-class}
\alias{NetworkAnalysedPerturbationData-class}
\title{Data wrapper for analysed perturbation data using network diffusion in}
\description{
Class \code{NetworkAnalysedPerturbationData} is a wrapper for
 various objects of an analysis of a perturbation experiment done
 using network diffusion. See also \code{\link{diffuse}}.
}
\section{Slots}{

\describe{
\item{\code{intitialModel}}{the model that was provided for analysis}

\item{\code{graph}}{an igraph object that served for the diffusion process}
}}

