% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-ora.R
\docType{methods}
\name{ora}
\alias{ora}
\alias{ora,character,character-method}
\title{Do overrepresentation analysis based on the hypergeometric test}
\usage{
ora(hit.list, universe, db = c("kegg", "go"), ...)

\S4method{ora}{character,character}(hit.list, universe, db = c("kegg", "go"),
  ...)
}
\arguments{
\item{hit.list}{a list of identified hit genes}

\item{universe}{a list of all genes}

\item{db}{either run ORA on KEGG or on Gene Ontology}

\item{...}{additional parameters}
}
\description{
Do overrepresentation analysis based on the hypergeometric test
}
