% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_getters.R, R/class_analysed.R
\docType{methods}
\name{params}
\alias{params}
\alias{params,AbstractAnalysedPerturbationData-method}
\title{Getter for parameters used for analysis of perturbation data}
\usage{
params(obj)

\S4method{params}{AbstractAnalysedPerturbationData}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the underlying params}
}
\value{
returns a \code{list}.
}
\description{
Returns the parameters used in the analysis of a perturbation
 analysis. Parameters are for instance the significance level, the effect
 size or restart probability of a random walk.
}
\examples{
 data(rnaiscreen)
 ft <- hm(rnaiscreen)
 params(ft)
}
