% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysed.R
\docType{class}
\name{PerturbationAnalysis-class}
\alias{PerturbationAnalysis-class}
\title{Data wrapper for analysed perturbation data}
\description{
Abstract class \code{perturbation.analysed} is a wrapper for a
  \code{data.table} object
containing the perturbation data
}
\section{Slots}{

\describe{
\item{\code{.data}}{the perturbation data-set}

\item{\code{.inference}}{the method for inferenced that has been used}

\item{\code{.is.bootstrapped}}{boolean whether bootstrap intervals have been}

\item{\code{.params}}{list of some used parameters
created or not}
}}

