% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data.R
\docType{class}
\name{HM-data}
\alias{HM-data}
\title{Data wrapper for hierarchical model perturbation data}
\description{
Class \code{perturbation.hm.data} is a wrapper for a
\code{data.table} object containing the normalized data-set
}
