% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data.R
\docType{class}
\name{Raw-data}
\alias{Raw-data}
\title{Data wrapper for raw perturbation data}
\description{
Class \code{perturbation.raw.data} is a wrapper for a
\code{data.table} object containing the raw data-set
}
