% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data.R
\docType{class}
\name{PerturbationData-class}
\alias{PerturbationData-class}
\title{Data wrapper for a data set of a perturbation screen.}
\description{
Class \code{PerturbationData} wraps a data set derived
 from a genetic perturbation screen, e.g. using RNA interference or CRISPR.
 Class \code{PerturbationData} exposes getters for its members of the same
 name, but no setters, because the data should be treated as
 constant once set. The easiest way to construct a \code{PerturbationData}
 object is by first creating a \code{data.frame} and then calling \code{as}.
 See the examples to construct an object.
}
\section{Slots}{

\describe{
\item{\code{dataSet}}{the data set as a \code{data.table} object}
}}

\examples{
  df <- data.frame(Condition    = c("V1", "V2", "V3"),
                   Replicate    = c(1, 1, 1),
                   GeneSymbol   = c("TP52", "NegCtrl", "PosCtrl"),
                   Perturbation = c("P1", "P2", "P3"),
                   Readout      = c(123, 121, 12),
                   Control      = c(0, -1, 1))
  methods::as(df, "PerturbationData")

}
