% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbR-package.R
\docType{package}
\name{perturbR-package}
\alias{perturbR-package}
\title{perturbR}
\description{
Statistical analysis of high-throughput genetic perturbation screens
}
\author{
Simon Dirmeier | Computational Biology Group,
 ETH ZURICH | \email{simon.dirmeier@bsse.ethz.ch}
}
\keyword{package}
