% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plate-data}
\alias{plate}
\alias{plate-data}
\title{plate}
\format{A data.table with 18 columns, each describing a necessary feature of a \code{svd.raw} data-set}
\usage{
data(plate)
}
\description{
Example \code{svd.raw} data-set consisting of a single plate with RANDOM data.
}
\author{
Simon Dirmeier | Computational Biology Group, ETH ZURICH | \email{simon.dirmeier@bsse.ethz.ch}
}
\keyword{data}
\keyword{datasets}

