% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plate_matrix.R
\name{plate.matrix}
\alias{plate.matrix}
\title{Get the plate matrix (plus control indexes) from an object}
\usage{
plate.matrix(obj, ...)
}
\arguments{
\item{obj}{the object for which you want to have the readout matrix}

\item{...}{additional params}
}
\description{
Get the plate matrix (plus control indexes) from an object
}
