% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.R
\name{plates}
\alias{plates}
\title{Get the plates of a data-set}
\usage{
plates(obj, ...)
}
\arguments{
\item{obj}{an object for which the plates are going to be retrieved}

\item{...}{additional params}
}
\description{
Get the plates of a data-set
}

