% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prioritized.R
\name{plot.effect.matrix}
\alias{plot.effect.matrix}
\title{Plot the effects matrix of a PMM analysis}
\usage{
plot.effect.matrix(x, ...)
}
\arguments{
\item{x}{an svd.prioritized.pmm object}

\item{...}{additional parameters}
}
\description{
Plot the effects matrix of a PMM analysis
}

