% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_analysed.R
\name{plot.knockdown.lmm.analysed}
\alias{plot.knockdown.lmm.analysed}
\title{Plot a \code{knockdown.lmm.analysed}  object}
\usage{
\method{plot}{knockdown.lmm.analysed}(x, size = 10, ...)
}
\arguments{
\item{x}{the object to be plotted}

\item{size}{size of letters}

\item{...}{additional parameters}
}
\value{
returns a list of plots
}
\description{
Plot a \code{knockdown.lmm.analysed}  object
}
