% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_other.R
\name{plot.knockdown.quality}
\alias{plot.knockdown.quality}
\title{Plot a \code{knockdown.quality} object}
\usage{
\method{plot}{knockdown.quality}(x, axis.text.size = 12, ...)
}
\arguments{
\item{x}{the object to plot}

\item{axis.text.size}{size of axis text}

\item{...}{additional parameters to plot}
}
\value{
returns a plot object
}
\description{
Plot a \code{knockdown.quality} object on a 2D grid
}
