% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot.knockdown.raw.data}
\alias{plot.knockdown.raw.data}
\title{Plot a knockdown dataset}
\usage{
\method{plot}{knockdown.raw.data}(x, size = 10, ...)
}
\arguments{
\item{x}{the object to plot}

\item{size}{size of letters}

\item{...}{additional parameters}
}
\value{
returns a plot object
}
\description{
Plot a knockdown dataset
}
