% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_other.R
\name{plot.knockdown.replicate}
\alias{plot.knockdown.replicate}
\title{Plot two \code{knockdown.replicate} objects}
\usage{
\method{plot}{knockdown.replicate}(x, y, method = c("Scatterplot", "QQ-plot"),
  size = 10, ...)
}
\arguments{
\item{x}{a \code{knockdown.replicate} object}

\item{y}{another \code{knockdown.replicate} object}

\item{method}{either \code{Scatterplot} or \code{QQ-plot}}

\item{size}{size of text}

\item{...}{additional parameters}
}
\value{
returns a plot object
}
\description{
Scatter two \code{knockdown.replicate} objects against
 each other
}
