% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diffusion.R
\name{plot.knockout.diffusion.analysed}
\alias{plot.knockout.diffusion.analysed}
\title{Plot a \code{knockout.diffusion.analysed} object}
\usage{
\method{plot}{knockout.diffusion.analysed}(x, graph.size = 20,
  show.node.labels = FALSE, ...)
}
\arguments{
\item{x}{a \code{knockout.diffusion.analysed} object}

\item{graph.size}{approximate numer of nodes}

\item{show.node.labels}{\code{logical} if gene names are shown or note}

\item{...}{addutional params}
}
\value{
returns a plot object
}
\description{
Plot a \code{knockout.diffusion.analysed} object
}
