% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot.knockout.raw.data}
\alias{plot.knockout.raw.data}
\title{Plot a knockout dataset}
\usage{
\method{plot}{knockout.raw.data}(x, size = 10, ...)
}
\arguments{
\item{x}{the object to plot}

\item{size}{size of letters}

\item{...}{additional parameters}
}
\description{
Plot a knockout dataset
}
