% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_analysed.R
\name{plot.knockout.tstatistic.analysed}
\alias{plot.knockout.tstatistic.analysed}
\title{Plot a \code{knockout.tstatistic.analysed} object}
\usage{
\method{plot}{knockout.tstatistic.analysed}(x, size = 10, ...)
}
\arguments{
\item{x}{the object to plot}

\item{size}{size of the text}

\item{...}{additional parameters}
}
\value{
returns a plot object
}
\description{
Plot a \code{knockout.tstatistic.analysed} object
}
