% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diffusion.R
\name{plot.NetworkAnalysedPerturbationData}
\alias{plot.NetworkAnalysedPerturbationData}
\title{Plot a \code{NetworkAnalysedPerturbationData} object}
\usage{
\method{plot}{NetworkAnalysedPerturbationData}(x, cnt = 10, ...)
}
\arguments{
\item{x}{a \code{NetworkAnalysedPerturbationData} object}

\item{cnt}{number of genes to be shown}

\item{...}{additional parameters}
}
\value{
returns a table if the first \code{cnt} highest ranked genes
}
\description{
Creates a table of the gene ranking of a
\code{NetworkAnalysedPerturbationData} object
}
