% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diffusion.R
\name{plot.perturbation.diffusion.analysed}
\alias{plot.perturbation.diffusion.analysed}
\title{Plot a \code{perturbation.mrw.diffusion.analysed} object}
\usage{
\method{plot}{perturbation.diffusion.analysed}(x, cnt = 10, ...)
}
\arguments{
\item{x}{a \code{perturbation.diffusion.analysed} object}

\item{cnt}{number of genes to be shown}

\item{...}{additional params}
}
\value{
returns a plot object
}
\description{
Plots a \code{perturbation.mrw.diffusion.analysed} object by
iteratively expanding the neighbors of the hits form the diffusion
}
