% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_analysed.R
\name{plot.perturbation.hyper.analysed}
\alias{plot.perturbation.hyper.analysed}
\title{Plot a \code{perturbation.hyper.analysed} object}
\usage{
\method{plot}{perturbation.hyper.analysed}(x, size = 10, ...)
}
\arguments{
\item{x}{the object to plot}

\item{size}{size of the text}

\item{...}{additional parameters}
}
\value{
returns a plot object
}
\description{
Plot a \code{perturbation.hyper.analysed} object
}
