% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot.perturbation.normalized.data}
\alias{plot.perturbation.normalized.data}
\title{Plot a perturbation data-set}
\usage{
\method{plot}{perturbation.normalized.data}(x, size, ...)
}
\arguments{
\item{x}{the object to plot}

\item{size}{size of letters}

\item{...}{additional parameters}
}
\value{
returns a plot object
}
\description{
Plot a perturbation data-set
}
