% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot.PerturbationData}
\alias{plot.PerturbationData}
\title{Plot perturbation data}
\usage{
\method{plot}{PerturbationData}(x, size = 10, ...)
}
\arguments{
\item{x}{the object to plot}

\item{size}{size of letters}

\item{...}{additional parameters}
}
\value{
returns a plot object
}
\description{
Creates a barplot of replicate and gene counts of a
 \code{PerturbationData} object.
}
