% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Pre-processing routine for data normalization, summarization and removal of low viability siRNAs.}
\usage{
preprocess(obj, normalize = c("log", "robust-z.score"),
  normalize.viability = F, rm.cytotoxic = NULL, rm.outlier.wells = c(NA,
  "quantile"), drop = T, ...)
}
\arguments{
\item{obj}{the data.table to be analysed}

\item{normalize}{a vector with normalization methods
\itemize{
 \item{\emph{log} }{ does a log transformation}
 \item{\emph{poc} }{ divides by the mean/median of the control}
 \item{\emph{loess} }{ fits a local regression model to correct for cell numbers per well}
 \item{\emph{b.score} }{ uses Tukey's median polish to correct for row and column effects}
 \item{\emph{z.score} }{ standardizes a plate}
 \item{\emph{robust-z.score} }{ standardizes a plate using the median instead of mean and the MAD instead if the standard deviation}
 \item{\emph{background} }{ substracts mean/median of some row/column or all NA genes}
 \item{\emph{qq} }{ quantile-quantile normalization on replicates}
}}

\item{normalize.viability}{boolean flag if the viability should also be removed}

\item{rm.cytotoxic}{gene name of neutral gene against which viability is testet (e.g. \emph{Scrambled} or \emph{PIK3CA})}

\item{rm.outlier.wells}{remove wells that have a extreme number of cells (outliers)}

\item{drop}{boolean if rows that are not needed should be dropped}

\item{...}{additional parameters
\itemize{
 \item{\emph{method} }{ either \emph{mean} or \emph{median} for diverse summarizations/..}
 \item{\emph{drop} }{ boolean flag if useless columns should be dropped or not}
 \item{\emph{z.score.level} }{ either \emph{plate} or  \emph{control}, depending on which the z-scores should be calculated}
 \item{\emph{z.score.ctrl} }{ gene name for which plate is normalized to when level=control}
 \item{\emph{poc.ctrl} }{ gene name for which plate is normalized to when poc is used}
 \item{\emph{background.row} }{ row idx that is used for background correction}
 \item{\emph{background.column} }{ column idx that is used for background correction}
 \item{\emph{outlier.well.range} }{ vector of lower and upper bound for excluded cells (e.g.: c(150,500) or c(.05, .95) depending on your choice of rm.outlier.wells) }
}}
}
\description{
Pre-processing routine for data normalization, summarization and removal of low viability siRNAs.
}

