% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{prioritize}
\alias{prioritize}
\title{Select hits from an analyzed data set}
\usage{
prioritize(obj, ...)
}
\arguments{
\item{obj}{the data.table to be called}

\item{...}{additional parameters
 \itemize{
 \item{\emph{hit.ratio} }{ the ratio of siRNAs hits such that a gene counts as hit}
 \item{\emph{readout.threshold} }{ the mean readout threshold for all the siRNAs}
 \item{\emph{p.value.threshold} }{ p.value threshold for hit selection}
 \item{\emph{fdr.threshold} }{ fdr.threshold}
}}
}
\description{
Select hits from an analyzed data set
}

