% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-quality.R
\name{quality}
\alias{quality}
\title{Calculates per-plate, per-replicate and screen quality scores}
\usage{
quality(obj, ...)
}
\arguments{
\item{obj}{the object for which quality scores are calculates}

\item{...}{additional parameters}
}
\description{
Calculates per-plate, per-replicate and screen quality scores
}
