% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_combine.R
\name{rbind.PerturbationData}
\alias{rbind.PerturbationData}
\title{Bind multiple perturbation data sets together by row}
\usage{
\method{rbind}{PerturbationData}(...)
}
\arguments{
\item{...}{variable number of \code{PerturbationData} objects}
}
\value{
returns a combined object of class \code{PerturbationData}
}
\description{
Binds multiple \code{PerturbationData} objects together by row.
}
\examples{
  data(rnaiscreen)
  rbind(rnaiscreen, rnaiscreen)

}
