% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_readout_matrix.R
\name{readout.matrix}
\alias{readout.matrix}
\title{Get the readout matrix (plus control indexes) from an object}
\usage{
readout.matrix(obj, ...)
}
\arguments{
\item{obj}{the object for which you want to have the readout matrix}

\item{...}{additional params}
}
\description{
Get the readout matrix (plus control indexes) from an object
}
