% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{replicates}
\alias{replicates}
\title{Get the replicates of a data-set}
\usage{
replicates(obj, ...)
}
\arguments{
\item{obj}{an object for which the replicates should be retrieved}

\item{...}{additional params}
}
\description{
Get the replicates of a data-set
}

