% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaiscreen}
\alias{rnaiscreen}
\title{A sample pan-pathogenic RNAi dataset}
\format{A \code{knockout.data} object containing a data.table with 18
 columns, each describing a necessary feature}
\usage{
data(rnaiscreen)
}
\description{
Example \code{knockout.data} data-set consisting of two-viral screen with
 random data.
}
\details{
\itemize{
    \item Virus  \code{character} names of the viruses
    \item Replicate  \code{integer} replicate numer
    \item Plate  \code{integer} plate index
    \item RowIdx  \code{integer} row index of the well on the plate
    \item ColIdx  \code{integer} column index of the well on the plate
    \item GeneSymbol  \code{character} HugoSymbol of a gene
    \item Entrez  \code{integer} entrez ID
    \item ReadoutType  \code{character} readouttype,
      such as `GFP` or `Luciferase`
    \item Control  \code{integer} coding of controls.
      `-1` for negative control, `1` for positive control, `0` for regular
       sample
    \item Library  \code{character} library type, such as `Ambion`
    \item siRNAIDs  \code{character} sirna identifier
    \item Screen  \code{character} identifier,
      for example `Kinome` or `Genome`
    \item Cell  \code{character} cell type, such as `Huh7.5`
    \item ScreenType \code{character} screen type, such as `E/R` for
       entry/replication
    \item Design  \code{character} design of the library, e.g. `pooled`
    \item Readout  \code{numeric} readout value
    \item ReadoutClass  \code{character} class of the readout, such as
      `Readout` or `Viability`
    \item NumCells  \code{integer} number of measured cells per well
 }
}
\author{
Simon Dirmeier | Computational Biology Group, ETH ZURICH |
 \email{simon.dirmeier@bsse.ethz.ch}
}
\keyword{data}
\keyword{datasets}
