% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_select.R
\name{select}
\alias{select}
\title{Select columns of a perturbation data set}
\usage{
select(obj, ...)
}
\arguments{
\item{obj}{the object of which columns should be selected}

\item{...}{variable number of unquoted columns names from \code{obj} to
subset from. Negative column names will be deselected. Wraps around
\code{dplyr::select}.}
}
\value{
returns a \code{tibble} with the specified columns
}
\description{
Takes a perturbation data set and selects columns by some
 criterion. The filtered object will have class \code{data.table}.
}
\examples{
 data(rnaiscreen)
 perturbatr::select(rnaiscreen, GeneSymbol, Readout)
}
