% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_lmm_model_data.R
\name{set.hm.model.data}
\alias{set.hm.model.data}
\title{Create model data for an hierarchical model}
\usage{
set.hm.model.data(obj, drop = TRUE, ignore = 1, weights = NULL,
  rel.mat.path = NULL)
}
\arguments{
\item{obj}{an object for which hm model.data is created}

\item{drop}{decide if genes that are not found in every virus should
be dropped}

\item{ignore}{ignore siRNAS that are only found \code{ignore} many times}

\item{weights}{weights to set for the siRNAs}

\item{rel.mat.path}{target-relation matrix (TODO)}
}
\value{
returns an \code{perturbation.hm.data} object
}
\description{
Create model data for an hierarchical model
}
