% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_lmm_model_data.R
\name{set.lmm.model.data}
\alias{set.lmm.model.data}
\title{Create model data for an LMM}
\usage{
set.lmm.model.data(obj, drop = T, ignore = 1, weights = NULL,
  rel.mat.path = NULL)
}
\arguments{
\item{obj}{an object for which LMM model.data is created}

\item{drop}{decide if genes that are not found in every virus should
be dropped}

\item{ignore}{ignore siRNAS that are only found \code{ignore} many times}

\item{weights}{weights to set for the siRNAs}

\item{rel.mat.path}{target-relation matrix (TODO)}
}
\description{
Create model data for an LMM
}
