% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_lmm_model_data.R
\docType{methods}
\name{setModelData}
\alias{setModelData}
\alias{setModelData,PerturbationData-method}
\title{Create model data for an hierarchical model}
\usage{
setModelData(obj, drop = TRUE, weights = NULL)

\S4method{setModelData}{PerturbationData}(obj, drop = TRUE, weights = NULL)
}
\arguments{
\item{obj}{an data set}

\item{drop}{boolean if genes that are not found in every Condition should
be dropped}

\item{weights}{a numeric vector used as weights for the single
perturbations}
}
\value{
returns an \code{PerturbationData} object
}
\description{
Create model data for an hierarchical model
}
\examples{
  data(rnaiscreen)
  rnaiscreen.normalized <- preprocess(rnaiscreen)
  setModelData(rnaiscreen.normalized)
}
