% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-effects.R
\name{show.effect.matrix}
\alias{show.effect.matrix}
\title{Plot the effects matrix of a PMM analysis}
\usage{
show.effect.matrix(x, ...)
}
\arguments{
\item{x}{an svd.prioritized.pmm object}

\item{...}{additional parameters}
}
\description{
Plot the effects matrix of a PMM analysis
}
