% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prioritized.R
\name{show.pathogen.hit.counts}
\alias{show.pathogen.hit.counts}
\title{Plot the hit counts of a PMM analysis for every virus as barplots}
\usage{
show.pathogen.hit.counts(x, ...)
}
\arguments{
\item{x}{an svd.prioritized.pmm object}

\item{...}{additional parameters}
}
\description{
Plot the hit counts of a PMM analysis for every virus as barplots
}

