% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{show.table}
\alias{show.table}
\title{Plot a table to the grid}
\usage{
show.table(obj, nrow = 10, colored = 2, main = "", ...)
}
\arguments{
\item{obj}{the object you want}
}
\description{
Plot a table to the grid
}

