% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\docType{methods}
\name{[,perturbation.plates,numeric,missing,missing-method}
\alias{[,perturbation.plates,numeric,missing,missing-method}
\title{Extract parts of an object}
\usage{
\S4method{[}{perturbation.plates,numeric,missing,missing}(x, i)
}
\arguments{
\item{x}{a \code{perturbation.plates} object}

\item{i}{index of the plate to subset}
}
\value{
returns a \code{perturbation.plate} object
}
\description{
Extract a single plate from a \code{perturbation.plates} object
}
