% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_tstatistic.R
\docType{methods}
\name{tstatistic}
\alias{tstatistic}
\alias{tstatistic,knockout.data-method}
\title{Calculate statistics based on the t-test to analyse the data.}
\usage{
tstatistic(obj, mu = c(0, "scrambled", "control"), padjust = c("BH",
  "bonferroni"), ...)

\S4method{tstatistic}{knockout.data}(obj, mu = c(0, "scrambled", "control"),
  padjust = c("BH", "bonferroni"), ...)
}
\arguments{
\item{obj}{the data to be analysed}

\item{mu}{side to which the mean of a siRNA is compared to}

\item{padjust}{multiple testing correction method}

\item{...}{additional params}
}
\description{
For this you should use a standardization before.
}
