% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sampler.R
\name{bootstrap}
\alias{bootstrap}
\title{Create a bootstrap sample from a data-set}
\usage{
bootstrap(obj, level = c("sirna", "condition"))
}
\arguments{
\item{obj}{the object which data should be bootstrapped}

\item{level}{boostrap either on perturbatinos or on conditions}
}
\value{
returns an object with boostrapped data
}
\description{
Create a bootstrap sample from a data-set
}
\examples{
  data(rnaiscreen)
  bootstrap(rnaiscreen)
}
