% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-chisq_statistic.R
\docType{methods}
\name{chisq.statistic}
\alias{chisq.statistic}
\alias{chisq.statistic,knockout.lmm.data-method}
\title{Calculate statistics based on the chi-square-distribution to analyse the
 data.}
\usage{
chisq.statistic(obj, padjust = c("BH", "bonferroni"), ...)

\S4method{chisq.statistic}{knockout.lmm.data}(obj, padjust = c("BH",
  "bonferroni"), ...)
}
\arguments{
\item{obj}{the data to be analysed}

\item{padjust}{multiple testing correction method}

\item{...}{additional params}
}
\description{
TODO
}
