% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correct}
\alias{correct}
\title{Off-target correct and rank by hit magnitude}
\usage{
correct(obj, path, drop, ...)
}
\arguments{
\item{obj}{a summarized data.table}

\item{path}{path (or file) to the target-relation matrices}

\item{drop}{drops all genes that are not found in every screen, i.e. a gene has to be found over all pathogens to be used}

\item{...}{additional arguments}
}
\description{
Off-target correct and rank by hit magnitude
}

