% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_diffuse.R
\docType{methods}
\name{diffuse}
\alias{diffuse}
\alias{diffuse,knockdown.lmm.analysed-method}
\title{Smooth the results from an analysis using network diffusion}
\usage{
diffuse(obj, path = NULL, graph = NULL, r = 0.5,
  delete.nodes.on.degree = 0, do.bootstrap = FALSE, ...)

\S4method{diffuse}{knockdown.lmm.analysed}(obj, path = NULL, graph = NULL,
  r = 0.5, delete.nodes.on.degree = 0, do.bootstrap = FALSE, ...)
}
\arguments{
\item{obj}{an analysed object}

\item{path}{path to the network file (if \code{graph} is \code{NULL})}

\item{graph}{an weighted adjacency matrix (if \code{path} is \code{NULL})}

\item{r}{restart probability of the random if method \code{mrw} is selected}

\item{delete.nodes.on.degree}{delete nodes from the graph with a degree of
less or equal than \code{delete.nodes.on.degree}}

\item{do.bootstrap}{run a diffusion on every bootstrap sample in case
bootstrap samples are available}

\item{...}{additional parameters}
}
\value{
returns a \code{knockdown.diffusion.analysed} object
}
\description{
TODO
}
\examples{
#load some data
data(rnaiscreen)
rnai.norm <- preprocess(rnaiscreen)
lmm.fit <- lmm(rnai.norm)
graph.file <- system.file("extdata", "graph_file.tsv", package="knockdown")

# do the diffusion using nearest neighbors
diff.knn <- diffuse(lmm.fit, method="m",
                    path=graph.file)

}
