% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_diffuse.R
\docType{methods}
\name{diffuse}
\alias{diffuse}
\alias{diffuse,perturbation.hm.analysed-method}
\title{Smooth the results from an analysis using network diffusion}
\usage{
diffuse(obj, path = NULL, graph = NULL, r = 0.5,
  delete.nodes.on.degree = 0, do.bootstrap = FALSE, ...)

\S4method{diffuse}{perturbation.hm.analysed}(obj, path = NULL, graph = NULL,
  r = 0.5, delete.nodes.on.degree = 0, do.bootstrap = FALSE, ...)
}
\arguments{
\item{obj}{an analysed object}

\item{path}{path to the network file (if \code{graph} is \code{NULL})}

\item{graph}{an weighted adjacency matrix (if \code{path} is \code{NULL})}

\item{r}{restart probability of the random if method \code{mrw} is selected}

\item{delete.nodes.on.degree}{delete nodes from the graph with a degree of
less or equal than \code{delete.nodes.on.degree}}

\item{do.bootstrap}{run a diffusion on every bootstrap sample in case
bootstrap samples are available}

\item{...}{additional parameters}
}
\value{
returns a \code{perturbation.diffusion.analysed} object
}
\description{
TODO
}
\examples{
\dontrun{
 data(rnaiscreen)
 rnaiscreen.normalized <- preprocess(rnaiscreen, normalize="robust-z.score")
 res                   <- hm(rnaiscreen.normalized, effect.size=0.01)
 graph.file <- system.file("extdata", "graph_file.tsv", package = "perturbatr")
 diffu      <- diffuse(res, path=graph.file, r=0.1)
}
}
