% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffuse.R
\name{diffuse}
\alias{diffuse}
\title{Extend the results from a <code>svd.prioritized</code> object by network diffusion.}
\usage{
diffuse(obj, method = c("neighbors", "mrw"), path = NULL, graph = NULL,
  r = 0.5, node.start.count = 25, search.depth = 5,
  delete.nodes.on.degree, ...)
}
\arguments{
\item{obj}{an \code{svd.prioritized} object}

\item{method}{method that should be used for diffusion
\itemize{
  \item{neighbors }{ just looks at the neighbors :)}
  \item{mrw }{ do a Markov random walk}
}}

\item{path}{path to the network file}

\item{graph}{an weighted adjacency matrix}

\item{r}{restart probability of the random if method \code{mrw} is selected}

\item{node.start.count}{number of nodes that are used to do the neighbors
search if method \code{neighbors} is selected
If the number of hits in \code{obj} exceeds \code{node.start.count},
then the elements with the highest absolute effects are chosen. If this
behaviour is not desired filter \code{obj} before}

\item{search.depth}{how deep should the neighbor search go if method \code{neighbors} is selected}

\item{delete.nodes.on.degree}{delete nodes from the graph with a degree of less that \code{delete.nodes.on.degree}}

\item{...}{additional parameters}
}
\description{
Extend the results from a <code>svd.prioritized</code> object by network diffusion.
}
