% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffuse.R
\name{diffuse}
\alias{diffuse}
\title{Extend the results from a <code>svd.prioritized</code> object by network diffusion.}
\usage{
diffuse(obj, method = c("neighbors", "mrw"), path, ...)
}
\arguments{
\item{obj}{an svd.data object}

\item{method}{method that should be used for diffusion
\itemize{
  \item{neighbors }{ just looks at the neighbors :)}
  \item{neighbors }{ just looks at the neighbors :)}
}}

\item{path}{path to the network}

\item{...}{additional parameters}
}
\description{
Extend the results from a <code>svd.prioritized</code> object by network diffusion.
}

