% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_effect_matrix.R
\name{effect.matrices}
\alias{effect.matrices}
\title{Calculcate the PMM-prioritized effect matrices for gene and pathogen-gene matrices}
\usage{
effect.matrices(obj, ...)
}
\arguments{
\item{obj}{the object to calculate the effect matrices for}

\item{...}{additional parameters}
}
\description{
Calculcate the PMM-prioritized effect matrices for gene and pathogen-gene matrices
}
