% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-filter.R
\docType{methods}
\name{filter}
\alias{filter}
\title{Filter the rows of a knockout data set}
\usage{
filter(obj, ...)
}
\arguments{
\item{obj}{the object to be filtered}

\item{...}{additional parameters}
}
\description{
\code{lmm} TODO
}
