% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_filter.R
\name{filter}
\alias{filter}
\title{Filter the rows of a perturbation data set}
\usage{
filter(obj, ...)
}
\arguments{
\item{obj}{the object to be filtered}

\item{...}{variable number of logical predicates in terms of the column
names in \code{obj}. Multiple predicates will be combined with a logical
'and'. Rows where all conditions are met will be kept. The column names do
 not need to be quoted. Wraps around
\code{dplyr::filter}.}
}
\value{
returns an object of the same type filtered by some criteria
}
\description{
Takes a perturbation data set and filters the rows by some
 criterion. The filtered object will have the same type as the previous
 object.
}
\examples{
 data(rnaiscreen)
 flt.dat <- filter(rnaiscreen, Condition=="V1")
 flt.dat <- filter(rnaiscreen, Condition=="V1", RowIdx==1)

}
