% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_getters.R, R/class_analysed.R
\docType{methods}
\name{graph}
\alias{graph}
\alias{graph,NetworkAnalysedPerturbationData-method}
\title{Getter for graph used for network diffusion}
\usage{
graph(obj)

\S4method{graph}{NetworkAnalysedPerturbationData}(obj)
}
\arguments{
\item{obj}{the object for which you want to extract the underlying graph}
}
\value{
returns a \code{igraph} object
}
\description{
Returns the graph that has been used for analysis using network
 diffusion.
}
\examples{
 \dontrun{
  data(rnaiscreen)
  graph.file <- system.file("extdata", "graph_file.tsv", package = "perturbatr")
  ft <- hm(rnaiscreen)
  diffu <- diffuse(ft, path=graph.file, r=0.1)
  graph(diffu)
 }
}
