% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_lmm.R
\docType{methods}
\name{hm}
\alias{hm}
\alias{hm,perturbation.normalized.data-method}
\alias{hm,perturbation.hm.data-method}
\title{Jointly analyse multiple genetic perturbation screens}
\usage{
hm(obj, drop = TRUE, weights = NULL, rel.mat.path = NULL,
  bootstrap.cnt = 0, ignore = 1, effect.size = 0.05,
  qval.threshold = 0.2)

\S4method{hm}{perturbation.normalized.data}(obj, drop = TRUE,
  weights = NULL, rel.mat.path = NULL, bootstrap.cnt = 0, ignore = 1,
  effect.size = 0.05, qval.threshold = 0.2)

\S4method{hm}{perturbation.hm.data}(obj, drop = TRUE, weights = NULL,
  rel.mat.path = NULL, bootstrap.cnt = 0, ignore = 1,
  effect.size = 0.01, qval.threshold = 0.2)
}
\arguments{
\item{obj}{an svd.data object}

\item{drop}{boolean flag if all entries should be dropped that are not
found in every Condition}

\item{weights}{a list of weights}

\item{rel.mat.path}{the (optional) path to a target relation matrix that is
going to be used for}

\item{bootstrap.cnt}{the number of bootstrap runs you want to do in order
to estimate a significance level for the gene effects}

\item{ignore}{ignore siRNAs that have been seen only once per group}

\item{effect.size}{the relative effect size used for hit prioritization}

\item{qval.threshold}{the q-value threshold used for hit prioritization
if bootstrap.cnt is set}
}
\value{
returns a \code{perturbation.hm.analysed} object
}
\description{
Analyse multiple different genetic perturbation screens at once
 using a hierarchical model. The model estimates general relative effect
 sizes for genes across all experiments. This could for instance be a
 pan-pathogenic host factor, i.e. a gene that decisively impacts the
 life-cycle of multiple pathogens.
}
