% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysed.R
\docType{class}
\name{HMAnalysedPerturbationData-class}
\alias{HMAnalysedPerturbationData-class}
\title{Data wrapper for analysed perturbation data using a hierarchical model}
\description{
Class \code{HMAnalysedPerturbationData} is a wrapper for
 various objects of an analysis of a perturbation experiment done
 using a hierarchical model.
}
\section{Slots}{

\describe{
\item{\code{geneEffects}}{the estimated effect sizes for genes}

\item{\code{geneHits}}{prioritized genes}

\item{\code{nestedGeneEffects}}{the estimated effect sizes for genes on a
viral level}

\item{\code{nestedGeneHits}}{nested prioritized genes}

\item{\code{modelFit}}{the fitted model}
}}

