% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysed.R
\docType{class}
\name{HMAnalysedPerturbationData-class}
\alias{HMAnalysedPerturbationData-class}
\title{Data wrapper for analysed perturbation data using a hierarchical model}
\description{
Class \code{HMAnalysedPerturbationData} is a wrapper for
 various objects of an analysis of a perturbation experiment done
 using a hierarchical model. Class \code{HMAnalysedPerturbationData} exposes
 getters for its members of the same name, but no setters, because the data
 should be treated as
 constant once set. Objects of class \code{HMAnalysedPerturbationData}
 do not need to be
 constructed manually but are returned from calling \code{\link{hm}} (see
 the examples).
}
\section{Slots}{

\describe{
\item{\code{nestedGeneEffects}}{the estimated effect sizes for genes on a
viral level}

\item{\code{modelFit}}{the fitted model}
}}

\examples{
 data(rnaiscreen)
 res <- hm(rnaiscreen, effect.size=0.01)
 class(res)
}
