% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_knockout_analysed.R
\docType{class}
\name{HyperAnalysis-class}
\alias{HyperAnalysis-class}
\title{Data wrapper for analysed knockout data using a standard hypothesis test}
\description{
Class \code{knockout.lmm.analysed} is a wrapper for a
  \code{data.table} object containing the knockout data
}
\section{Slots}{

\describe{
\item{\code{.gene.effects}}{the estimated effect sizes for genes}

\item{\code{.gene.pathogen.effects}}{the estimated effect sizes for genes on a
viral level}

\item{\code{.infectivity.effects}}{the  estimated effect sizes for different
infectivity levels}

\item{\code{.gene.hits}}{prioritized genes}

\item{\code{.gene.pathogen.hits}}{prioritized genes on a viral level}

\item{\code{.model.fit}}{the fitted model with gene fdrs and gene-pathogen
fdrs}
}}

