% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{knockout.data-class}
\alias{knockout.data-class}
\alias{knockout.data}
\title{Data wrapper for knockout data}
\description{
Data wrapper for knockout data
}
\section{Slots}{

\describe{
\item{\code{data}}{the RNAi data-set}
}}

