% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-knockout_data.R
\docType{class}
\name{knockout.lmm.data-class}
\alias{knockout.lmm.data-class}
\alias{knockout.lmm.data}
\title{Data wrapper for knockout linear mixed model data.}
\description{
Class \code{knockout.lmm.data} is a wrapper the data used by
 LMM.
}
\section{Slots}{

\describe{
\item{\code{.data}}{the knockout data-set}
}}

