% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-knockout_data.R
\docType{class}
\name{knockout.data.raw-class}
\alias{knockout.data.raw-class}
\alias{knockout.data.raw}
\title{Data wrapper for unnormalized knockout data}
\description{
Class \code{knockout.data.raw} is a wrapper for a \code{data.table} object
containing the unnormalized knockout data
}
\section{Slots}{

\describe{
\item{\code{.data}}{the unnormalized knockout data-set}
}}

