% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm}
\alias{lmm}
\title{Fit an LMM to the data and calculate local false discovery rates.}
\usage{
lmm(obj, drop = T, weights = NULL, rel.mat.path = NULL,
  bootstrap.cnt = F, ignore = 1, ...)
}
\arguments{
\item{obj}{an svd.data object}

\item{drop}{boolean flag if all entries should be dropped that are not found in every virus}

\item{weights}{a list of weights}

\item{rel.mat.path}{the (optional) path to a target relation matrix that is going to be used for}

\item{bootstrap.cnt}{the number of loocv runs you want to do in order to estimate a significance level for the gene effects}

\item{ignore}{ignore siRNAs that have been seen only once per group}

\item{...}{additional parameters}
}
\description{
Fit an LMM to the data and calculate local false discovery rates.
}
