% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm}
\alias{lmm}
\title{Fit an LMM to the data and calculate local false discovery rates.}
\usage{
lmm(obj, th = 0.1, ...)
}
\arguments{
\item{obj}{an svd.data object}

\item{th}{the threshold for the local false discovery rate}

\item{...}{additional parameters}
}
\description{
Fit an LMM to the data and calculate local false discovery rates.
}

