% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm}
\alias{lmm}
\title{Fit an LMM to the data and calculate local false discovery rates.}
\usage{
lmm(obj, drop = T, ...)
}
\arguments{
\item{obj}{an svd.data object}

\item{drop}{boolean flag if all entries should be dropped that are not found in every virus}

\item{...}{additional parameters
\itemize{
 \item{ignore }{ remove sirnas that have been found less than \code{ignore} times}
}}
}
\description{
Fit an LMM to the data and calculate local false discovery rates.
}

