% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysed.R
\docType{class}
\name{NetworkAnalysedPerturbationData-class}
\alias{NetworkAnalysedPerturbationData-class}
\title{Data wrapper for analysed perturbation data using network diffusion}
\description{
Class \code{NetworkAnalysedPerturbationData} is a wrapper for
 various objects of an analysis of a perturbation experiment done
 using network diffusion. Class \code{NetworkAnalysedPerturbationData}
exposes getters for its members of the same name,
 but no setters, because the data should be treated as
 constant once set. Objects of class \code{NetworkAnalysedPerturbationData}
 do not need to be
 constructed manually but are returned from calling \code{\link{diffuse}}
 (see the examples).
}
\section{Slots}{

\describe{
\item{\code{initialModel}}{the model that was provided for analysis}

\item{\code{graph}}{an igraph object that served for the diffusion process}
}}

\examples{
 data(rnaiscreen)
 hm.fit <- hm(rnaiscreen, effect.size=0.01)
 graph.file <- system.file("extdata", "graph_file.tsv",
                                       package = "perturbatr")
 res <- diffuse(hm.fit, path=graph.file, r=1)
}
