% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data.R
\docType{class}
\name{PerturbationData-class}
\alias{PerturbationData-class}
\title{Data wrapper for a data set of a perturbation screen.}
\description{
Class \code{PerturbationData} wraps a data set derived
 from a genetic perturbation screen, e.g. using RNA interference or CRISPR.
}
\section{Slots}{

\describe{
\item{\code{dataSet}}{the data set as a \code{data.table} object}

\item{\code{dataType}}{the type that describes your data best, for instance
\code{raw} or \code{normalized}}
}}

