% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbatr-package.R
\docType{package}
\name{perturbatr-package}
\alias{perturbatr-package}
\title{perturbatr}
\description{
Statistical analysis of high-throughput genetic perturbation screens
}
\author{
Simon Dirmeier | Computational Biology Group,
 ETH ZURICH | \email{simon.dirmeier@web.de}
}
\keyword{package}
