% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plates.R
\name{plates}
\alias{plates}
\title{Get the plates of a data-set}
\usage{
plates(obj, ...)
}
\arguments{
\item{obj}{an object for which the plates are going to be retrieved}

\item{...}{additional params}
}
\value{
returns a \code{knockout.plates} object
}
\description{
Get the plates of a data-set
}
\examples{
 data(rnaiscreen)
 plates <- plates(rnaiscreen)
}
