% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prioritized.R
\name{plot.knockout.lmm.analysed}
\alias{plot.knockout.lmm.analysed}
\title{Plot an \code{knockout.lmm.analysed} object}
\usage{
\method{plot}{knockout.lmm.analysed}(x, ...)
}
\arguments{
\item{x}{the object to be plotted}

\item{...}{additional parameters}
}
\description{
The method returns three individual plots as a list.
}
