% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot.knockout.normalized.data}
\alias{plot.knockout.normalized.data}
\title{Plot a knockout data-set}
\usage{
\method{plot}{knockout.normalized.data}(x, ...)
}
\arguments{
\item{x}{the object to plot}

\item{...}{additional parameters}
}
\description{
Plot a knockout data-set
}
