% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_other.R
\name{plot.knockout.plate}
\alias{plot.knockout.plate}
\title{Plot a \code{knockout.plate}}
\usage{
\method{plot}{knockout.plate}(x, y, show.controls = FALSE,
  show.gene.names = FALSE, ylab = "Row idx", xlab = "Column idx",
  main = "", ...)
}
\description{
@export
}
