% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_other.R
\name{plot.knockout.plate}
\alias{plot.knockout.plate}
\title{Plot a \code{knockout.plate} object}
\usage{
\method{plot}{knockout.plate}(x, show.controls = FALSE,
  show.gene.names = FALSE, ylab = "Row idx", xlab = "Column idx",
  main = "", axis.text.size = 12, gene.text.size = 3, ...)
}
\arguments{
\item{x}{the object to plot}

\item{show.controls}{show which wells are controls}

\item{show.gene.names}{show the gene names for every well}

\item{ylab}{a title for the ylab}

\item{xlab}{a title for the xlab}

\item{main}{the title of the plot}

\item{axis.text.size}{text size of axis labels}

\item{gene.text.size}{the size of the gene names within each well}

\item{...}{additional params}
}
\description{
Plot a \code{knockout.plate} object on a 2D grid
}
