% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_other.R
\name{plot.knockout.quality}
\alias{plot.knockout.quality}
\title{Plot a \code{knockout.quality} object}
\usage{
\method{plot}{knockout.quality}(x, ...)
}
\description{
Plot a \code{knockout.quality} object on a 2D grid
}
