% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\docType{methods}
\name{preprocess}
\alias{preprocess}
\alias{preprocess,knockout.data-method}
\alias{preprocess,knockout.data.raw-method}
\title{Pre-processing routine for data normalization, summarization and removal
of low viability siRNAs.}
\usage{
preprocess(obj, normalize = c("log", "robust-z.score"),
  normalize.viability = F, rm.cytotoxic = NULL,
  rm.outlier.wells = c("none", "quantile", "absolute"),
  z.score.level = c("plate", "control"), z.score.mu = "scrambled",
  poc.ctrl = "scrambled", background.row = NULL, background.column = NULL,
  outlier.well.range = c(0.05, 0.95), summarization = c("mean", "median"),
  drop = T)

\S4method{preprocess}{knockout.data}(obj, normalize = c("log",
  "robust-z.score"), normalize.viability = F, rm.cytotoxic = NULL,
  rm.outlier.wells = c("none", "quantile", "absolute"),
  z.score.level = c("plate", "control"), z.score.mu = "scrambled",
  poc.ctrl = "scrambled", background.row = NULL, background.column = NULL,
  outlier.well.range = c(0.05, 0.95), summarization = c("mean", "median"),
  drop = T)
}
\arguments{
\item{obj}{the object to be normalized}

\item{normalize}{a vector with normalization methods
\itemize{
 \item{\emph{log} }{ does a log transformation}
 \item{\emph{poc} }{ divides by the mean/median of the control}
 \item{\emph{loess} }{ fits a local regression model to correct
                       for cell numbers per well}
 \item{\emph{b.score} }{ uses Tukey's median polish to correct
                         for row and column effects}
 \item{\emph{z.score} }{ standardizes a plate}
 \item{\emph{robust-z.score} }{ standardizes a plate using the median
                                instead of mean and the MAD instead if the
                                standard deviation}
 \item{\emph{background} }{ substracts mean/median of some row/column or
                            all NA genes}
 \item{\emph{qq} }{ quantile-quantile normalization on replicates}
}}

\item{normalize.viability}{boolean flag if the viability should also be
normalized}

\item{rm.cytotoxic}{gene name of neutral gene against which viability is
testet (e.g. \emph{Scrambled} or \emph{PIK3CA}), etc.}

\item{rm.outlier.wells}{remove wells that have a extreme number of cells
(outliers), e.g. by taking quantiles or absolute numbers}

\item{z.score.level}{if z-scores are used for normalisation, choose either
'plate' or 'control' to calculate statistics for}

\item{z.score.mu}{the gene you want to compare to, e.g. 'scrambled', 'tp53'}

\item{poc.ctrl}{the control used for poc-normalization, e.g. 'scrambled'}

\item{background.row}{index of the row to correct backgorund against}

\item{background.column}{index of column to correct backgorund against}

\item{outlier.well.range}{outlier cells will be removed based on this range}

\item{summarization}{method of how to summarize screens (needed where? :/)}

\item{drop}{boolean if rows that are not needed should be dropped or
kept (e.g. if you want to check whether all worked out correctly)}
}
\description{
\code{preproces} normalizes the data for comparable phenotypes.
TODO
}
