% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_quality.R
\name{quality}
\alias{quality}
\title{Calculates per-plate, per-replicate and screen quality scores}
\usage{
quality(obj, ...)
}
\arguments{
\item{obj}{the object for which quality scores are calculates}

\item{...}{additional parameters}
}
\value{
returns a \code{knockout.quality} object
}
\description{
Calculates per-plate, per-replicate and screen quality scores
}
\examples{
 data(rnaiscreen)
 qual <- quality(rnaiscreen)
}
