% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-combine.R
\docType{methods}
\name{rbind}
\alias{rbind}
\alias{rbind,knockout.data-method}
\title{Combine objects}
\usage{
rbind(...)

\S4method{rbind}{knockout.data}(...)
}
\arguments{
\item{...}{the objects to be combined}
}
\description{
Combine several objects to a single objects.
}
