% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_replicates.R
\name{replicates}
\alias{replicates}
\title{Get the replicates of a data-set}
\usage{
replicates(obj, ...)
}
\arguments{
\item{obj}{an object for which the replicates should be retrieved}

\item{...}{additional params}
}
\value{
returns a \code{knockout.replicates} object
}
\description{
Get the replicates of a data-set
}
\examples{
 data(rnaiscreen)
 rep.dat <- replicates(rnaiscreen)
}
