% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_select.R
\name{select}
\alias{select}
\title{Select columns of a knockout data set}
\usage{
select(obj, ...)
}
\arguments{
\item{obj}{the object of which columns should be selected}

\item{...}{additional parameters}
}
\description{
\code{lmm} TODO
}
