% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_select.R
\name{select}
\alias{select}
\title{Select columns of a perturbation data set}
\usage{
select(obj, ...)
}
\arguments{
\item{obj}{the object of which columns should be selected}

\item{...}{additional parameters}
}
\value{
returns a \code{data.table} with specified column
}
\description{
Takes a perturbation data set and selects columns by some
 criterion. The filtered object will have class \code{data.table}.
}
\examples{
 data(rnaiscreen)
 perturbatr::select(rnaiscreen, GeneSymbol, Readout)
}
