% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\docType{methods}
\name{[,knockout.replicates,numeric,missing,missing-method}
\alias{[,knockout.replicates,numeric,missing,missing-method}
\title{Extract parts of an object}
\usage{
\S4method{[}{knockout.replicates,numeric,missing,missing}(x, i)
}
\arguments{
\item{x}{a \code{knockout.replicates} object}

\item{i}{index of the plate to subset}
}
\value{
returns a \code{knockout.replicate} object
}
\description{
Extract a single replicate from a \code{knockout.replicates} object
}
