% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_tstatistic.R
\docType{methods}
\name{tstatistic}
\alias{tstatistic}
\alias{tstatistic,perturbation.normalized.data-method}
\alias{tstatistic,perturbation.data-method}
\title{Calculate statistics based on the t-test to analyse the data.}
\usage{
tstatistic(obj, mu = c(0, "scrambled", "control"), padjust = c("BH",
  "bonferroni"), hit.ratio = 0.5, effect.size = 0, pval.threshold = 0.05,
  qval.threshold = 1)

\S4method{tstatistic}{perturbation.normalized.data}(obj, mu = c(0,
  "scrambled", "control"), padjust = c("BH", "bonferroni"), hit.ratio = 0.5,
  effect.size = 0, pval.threshold = 0.05, qval.threshold = 1)
}
\arguments{
\item{obj}{the data to be analysed}

\item{mu}{side to which the mean of a siRNA is compared to}

\item{padjust}{multiple testing correction method}

\item{hit.ratio}{the ratio of siRNAs}

\item{effect.size}{the relative strength of a signal to count as a hit,
i. e. the biological significance of a gene/sirna}

\item{pval.threshold}{the significance level for a sirna/gene,
i. e. the statistical significance of a gene/sirna
to be counted as significant}

\item{qval.threshold}{the significance level of the multiple testing
corrected p-value. This should be set to an appropriate significance level
just like the \code{pval.threshold} as well.}
}
\value{
returns a \code{perturbation.tstatistic.analysed} object
}
\description{
For this you should use a standardization before.
}
\examples{
 data(rnaiscreen)
 v1.dat <- perturbatr::filter(rnaiscreen, Condition=="V1")
 v1.data.norm <- perturbatr::preprocess(v1.dat)
 v1.res <- tstatistic(v1.data.norm)
}
